Posible fix of static object overlaping
