+++ Zilon.Core/Zilon.Core/MapGenerators/IStaticObjectGenerationContext.cs
@@ -1,5 +1,4 @@
-﻿using System;
-
+﻿
 using Zilon.Core.Schemes;
 using Zilon.Core.Tactics;
 using Zilon.Core.World;
@@ -11,20 +10,5 @@ namespace Zilon.Core.MapGenerators
         IResourceDepositData ResourceDepositData { get; }
         ISectorSubScheme Scheme { get; }
         ISector Sector { get; }
-    }
-
-    public sealed class StaticObjectGenerationContext : IStaticObjectGenerationContext
-    {
-        public StaticObjectGenerationContext(ISector sector, ISectorSubScheme scheme,
-            IResourceDepositData resourceDepositData)
-        {
-            Sector = sector ?? throw new ArgumentNullException(nameof(sector));
-            Scheme = scheme ?? throw new ArgumentNullException(nameof(scheme));
-            ResourceDepositData = resourceDepositData ?? throw new ArgumentNullException(nameof(resourceDepositData));
-        }
-
-        public ISector Sector { get; }
-        public ISectorSubScheme Scheme { get; }
-        public IResourceDepositData ResourceDepositData { get; }
     }
 }
