+++ Zilon.Core/Zilon.TextClient/MainScreenHandler.cs
@@ -73,6 +73,7 @@ namespace Zilon.TextClient
                 Console.WriteLine(UiResource.IdleCommandDescription);
                 Console.WriteLine(UiResource.DeadCommandDescription);
                 Console.WriteLine(UiResource.ExitCommandDescription);
+                Console.WriteLine(UiResource.TransitionCommandDescription);
 
                 Console.WriteLine($"{UiResource.InputPrompt}:");
                 var inputText = Console.ReadLine();
