+++ Zilon.Core/Zilon.TextClient/UiResource.resx
@@ -159,6 +159,9 @@
   <data name="StaticObjectNodeMarker" xml:space="preserve">
     <value>object</value>
   </data>
+  <data name="TransitionCommandDescription" xml:space="preserve">
+    <value>- "transit" to go to next level</value>
+  </data>
   <data name="TransitionNodeMarker" xml:space="preserve">
     <value>t</value>
   </data>
